+++ libs/notitia-templates/src/templates/utils/index.ts
@@ -1,0 +1,4 @@
+export { buildOrderConfirmationTemplate, ORDER_CONFIRMATION_STYLES } from './orderConfirmationBuilder';
+export type { OrderConfirmationTexts } from './orderConfirmationBuilder';
+export { buildOrderShippedTemplate, ORDER_SHIPPED_STYLES } from './orderShippedBuilder';
+export type { OrderShippedTexts } from './orderShippedBuilder';